Minor cleanup, adding license info
